:tada: Update layout from collectionview to tableview #2
