+++ SwiftyHome/Application/Extensions/UIKit/UITableView+.swift
@@ -1,13 +1,12 @@
 //
-//  UICollectionView+.swift
+//  UITableView+.swift
 //  SwiftyHome
 //
 //  Created by Yuto Mizutani on 2019/02/16.
 //  Copyright © 2019 Yuto Mizutani. All rights reserved.
 //
 
-import class UIKit.UICollectionViewCell
-import class UIKit.UICollectionView
+import UIKit
 import struct Foundation.IndexPath
 
 protocol ReusableView: class {
@@ -21,10 +20,21 @@ extension ReusableView {
 }
 
 extension UICollectionViewCell: ReusableView {}
+extension UITableViewCell: ReusableView {}
 
 extension UICollectionView {
     func dequeueReusableCell<T: UICollectionViewCell>(for indexPath: IndexPath) -> T {
         guard let cell = dequeueReusableCell(withReuseIdentifier: T.reuseIdentifier, for: indexPath) as? T else {
+            fatalError("Could not dequeue cell with identifier: \(T.reuseIdentifier)")
+        }
+
+        return cell
+    }
+}
+
+extension UITableView {
+    func dequeueReusableCell<T: UITableViewCell>(for indexPath: IndexPath) -> T {
+        guard let cell = dequeueReusableCell(withIdentifier: T.reuseIdentifier, for: indexPath) as? T else {
             fatalError("Could not dequeue cell with identifier: \(T.reuseIdentifier)")
         }
